refactor(feed): index blocks by u64, matching the store and the wire
`Log::get`, `Log::proof`, `Replica::block`, and `Source::{get,proof}` all took `usize` and
immediately did `index as u64` internally. Everything underneath is u64 — the store keys,
the accumulator, `Head::len`, and the wire's `GetBlock { index }` — so the `usize` surface
was an API accident that pushed a *narrowing* cast onto every caller holding a real feed
index. On a 32-bit target those casts wrap.

This wasn't only untidy. `sync::serve_feed` took the wire index (u64) and ran
`usize::try_from` purely to call `Source::get`/`proof`, so a 32-bit server answered `Absent`
for any index at or past 2^32 — refusing blocks it actually held, indistinguishably from
genuinely not having them. That narrowing is now gone rather than guarded.

It also retires a workaround from #55, landed yesterday: `run_mirror_window`'s pin scan had
to `usize::try_from(i)` and skip unaddressable indices, because a wrapped cast would have
tested some *other* block's bytes and pinned `i` on a decision that wasn't about `i`. With
the parameter widened there is nothing to guard, so the guard is deleted.

Net effect across the workspace: 19 narrowing `as usize` casts removed, none added. What
casts remain on these paths are the safe direction — a `Vec`-index `usize` widened to u64 at
a call site (`props.rs`, where `i` genuinely indexes a `Vec<Vec<u8>>`), or `len() as u64` in
tests. `len()` still returns `usize` and so still narrows internally; widening it is a
larger, separate question (it interacts with `is_empty`, arithmetic, and Rust's convention
for collection lengths), so it's deliberately not in scope here.

Gates: fmt, clippy -D warnings, cargo doc -D warnings, and all 39 test binaries pass.

+++ crates/feed/tests/soak.rs
@@ -36,9 +36,9 @@ fn a_windowed_seeder_stays_disk_bounded_as_the_author_grows() {
         let mut mirror = Replica::sparse(pk, author.head(), author.peak_nodes(), store.clone())
             .expect("peaks reproduce the head root");
         for i in start..len {
-            let proof = author.proof(i as usize).unwrap();
+            let proof = author.proof(i).unwrap();
             assert!(
-                mirror.ingest(i, author.get(i as usize).unwrap(), &proof),
+                mirror.ingest(i, author.get(i).unwrap(), &proof),
                 "window block {i} ingests"
             );
         }
@@ -51,7 +51,7 @@ fn a_windowed_seeder_stays_disk_bounded_as_the_author_grows() {
             "held window slides with the author (len {len})"
         );
         for i in start..len {
-            assert_eq!(mirror.block(i as usize), author.get(i as usize));
+            assert_eq!(mirror.block(i), author.get(i));
         }
 
         // Disk is bounded: the store holds only the window's blocks, never the whole feed.
